Deprecate serial protocol [ci skip]
After some research, I have decided that adding an Ethernet jack
and accompanying microcontroller to the cart is much easier than
writing a custom protocol and accompanying hardware for the NES
expansion port.

While this may increase the time it takes to develop NEScord, I
feel that in the long run it will be the best approach as it means
that all of our hardware can be in a single package to make
development costs cheaper.